Load CRDs for envtests from chart directory
Followup of #2145: given that `config` folder is going to be removed soon,
`internal/operator-controller/controllers/suite_test.go` loads CRDs from
`helm/olmv1/base/operator-controller/crd`.

Creation of `envtest.Environment` moved and consolidated into `test/utils.go`
so that it can be consumed by multiple test suites.
